centralized error handling added

+++ controllers/cards.js
@@ -1,56 +1,52 @@
 const { celebrate, Joi } = require('celebrate');
 const Card = require('../models/card');
+const NoAuth = require('../errors/no-auth');
+const BadRequest = require('../errors/bad-request');
+const NotFound = require('../errors/not-found');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate('user')
     .then((cards) => res.send({ data: cards }))
-    .catch(() => { res.status(500).send({ message: 'Ошибка!' }); });
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: 'Ошибка валидации' });
+    .then((card) => {
+      if (!card) {
+        throw new BadRequest('Ошибка валидации');
       }
-      res.status(500).send({ message: 'Ошибка!' });
-    });
+      res.send({ data: card });
+    })
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  const {ownerId} = req.user._id
+  const { ownerId } = req.user._id;
   Card.findById(cardId)
     .orFail(() => {
-      const error = new Error('Карточка по данному id не найдена');
-      error.statusCode = 404;
-      throw error;
+      throw new NotFound('Карточка по данному id не найдена');
     })
     .then((card) => {
+      if (!card) {
+        throw new BadRequest('Неверные данные');
+      }
       if (ownerId === card.owner._id) {
         Card.deleteOne(card).then(() => {
           res.send({ data: card });
         });
       } else {
-        res.status(401).send({ message: 'Недостаточно прав' });
+        throw new NoAuth('Недостаточно прав');
       }
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Неверные данные' });
-      } else if (err.statusCode === 404) {
-        res.status(err.statusCode).send({ message: err.message });
-      } else {
-        res.status(500).send({ message: err.message });
-      }
-    });
+    .catch(next);
 };
 
-module.exports.setLike = (req, res) => {
+module.exports.setLike = (req, res, next) => {
   const id = req.user._id;
   Card.findByIdAndUpdate(
     req.params.cardId,
@@ -58,40 +54,28 @@ module.exports.setLike = (req, res) => {
     { new: true },
   )
     .orFail(() => {
-      const error = new Error('Карточка по данному id не найдена');
-      error.statusCode = 404;
-      throw error;
+      throw new NotFound('Карточка по данному id не найдена');
     })
     .then((card) => {
+      if (!card) {
+        throw new BadRequest('Неверные данные');
+      }
       res.send({ data: card });
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Неверные данные' });
-      } else if (err.statusCode === 404) {
-        res.status(err.statusCode).send({ message: err.message });
-      } else {
-        res.status(500).send({ message: err.message });
-      }
-    });
+    .catch(next);
 };
 
-module.exports.removeLike = (req, res) => {
+module.exports.removeLike = (req, res, next) => {
   const id = req.user._id;
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: id } }, { new: true })
     .orFail(() => {
-      const error = new Error('Неверные данные');
-      error.statusCode = 404;
-      throw error;
+      throw new NotFound('Карточка по данному id не найдена');
     })
-    .then((card) => { res.send({ data: card }); })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Неверные данные' });
-      } else if (err.statusCode === 404) {
-        res.status(err.statusCode).send({ message: err.message });
-      } else {
-        res.status(500).send({ message: err.message });
+    .then((card) => {
+      if (!card) {
+        throw new BadRequest('Неверные данные');
       }
-    });
+      res.send({ data: card });
+    })
+    .catch(next);
 };
